Correct and clarify README.md
Made some adjustments to the headings so it's clear that the grid is
comprised of containers, rows and cols.  Moved the section I added
regarding the Bootstrap color system to the end so the change can be
demo'd once devs finish their labs.  Clarified and simplified some of
the language that seemed overly complicated.